Fix bugs in visualization

+++ BuildingMassEstimation/GLWidget3D.h
@@ -61,8 +61,8 @@ public:
 	void render();
 	void loadCGA(const std::string& cga_filename);
 	void generateTrainingData(const QString& cga_dir, const QString& out_dir, int numSamples, int image_width, int image_height, bool grayscale, bool centering, int cameraType, float cameraDistanceBase, float cameraHeight, int xrotMin, int xrotMax, int yrotMin, int yrotMax, int fovMin, int fovMax);
-	void generateTrainingDataWithoutAmgiousViewpoints(const QString& cga_dir, const QString& out_dir, int numSamples, int image_width, int image_height, bool grayscale, bool centering, float xrotMin, float xrotMax, float yrotMin, float yrotMax);
-	void visualizePredictedData(const QString& cga_dir, const QString& out_dir, int cameraType, float cameraDistanceBase, float cameraHeight, float xrotMin, float xrotMax, float yrotMin, float yrotMax, float fovMin, float fovMax);
+	void generateTrainingDataWithoutAmgiousViewpoints(const QString& cga_dir, const QString& out_dir, int numSamples, int image_width, int image_height, bool grayscale, bool centering, int xrotMin, int xrotMax, int yrotMin, int yrotMax);
+	void visualizePredictedData(const QString& cga_dir, const QString& out_dir, int cameraType, float cameraDistanceBase, float cameraHeight, int xrotMin, int xrotMax, int yrotMin, int yrotMax, int fovMin, int fovMax);
 	void runMCMC(const std::string& cga_filename, const std::string& target_filename, int numIterations);
 	void runMCMCAll(const std::string& cga_dir, int numIterations);
 	bool moveCenter(cv::Mat& img);
